examples/agent-runtime/agentic-frameworks/openai-agents-sdk/requirements.txt
# Novita Sandbox SDK
novita-sandbox>=1.1.0b1

# OpenAI SDK
openai>=1.50.0

# Utilities
python-dotenv>=1.0.0

